+++ scanner/phase_1_folders.go
@@ -129,6 +129,21 @@ type phaseFolders struct {
 	imageChangedFolders map[int][]imageChangedFolder
 }
 
+// imageChangedFolder records a folder whose image files changed during the scan, so the
+// affected albums/artists can be re-enqueued for artwork resolution at the end of phase 1.
+type imageChangedFolder struct {
+	id          string
+	path        string
+	artistImage bool
+}
+
+func scanArtworkItem(kind model.Kind, id string) model.ArtworkQueueItem {
+	return model.ArtworkQueueItem{
+		ItemKind: kind.Prefix(), ItemID: id, ImageType: model.ImageTypePrimary,
+		Priority: model.ArtworkPriorityScan,
+	}
+}
+
 func (p *phaseFolders) description() string {
 	return "Scan all libraries and import new/updated files"
 }
@@ -500,6 +515,74 @@ func (p *phaseFolders) logFolder(entry *folderEntry) (*folderEntry, error) {
 		"elapsed", entry.elapsed.Elapsed(), "tracksMissing", len(entry.missingTracks),
 		"tracksImported", len(entry.tracks), "library", entry.job.lib.Name, consts.Zwsp+"folder", entry.path)
 	return entry, nil
+}
+
+// enqueueArtworkForImageChanges re-enqueues artwork for entities affected by image-only folder
+// changes. Best-effort: failures are logged and never fail the scan.
+func (p *phaseFolders) enqueueArtworkForImageChanges() {
+	for _, job := range p.jobs {
+		folders := p.imageChangedFolders[job.lib.ID]
+		if len(folders) == 0 {
+			continue
+		}
+		items, err := p.collectImageChangeQueueItems(job.lib, folders)
+		if err != nil {
+			log.Warn(p.ctx, "Scanner: could not map image changes to artwork items", "lib", job.lib.Name, err)
+			continue
+		}
+		if len(items) == 0 {
+			continue
+		}
+		if err := p.ds.ArtworkQueue(p.ctx).Enqueue(items...); err != nil {
+			log.Warn(p.ctx, "Scanner: could not enqueue artwork for image changes", "lib", job.lib.Name, err)
+			continue
+		}
+		log.Debug(p.ctx, "Scanner: Enqueued artwork resolution for image changes", "lib", job.lib.Name,
+			"changedFolders", len(folders), "items", len(items))
+	}
+}
+
+func (p *phaseFolders) collectImageChangeQueueItems(lib model.Library, folders []imageChangedFolder) ([]model.ArtworkQueueItem, error) {
+	folderIDs := make([]string, len(folders))
+	var artistFolderPaths []string
+	for i, f := range folders {
+		folderIDs[i] = f.id
+		if f.artistImage {
+			artistFolderPaths = append(artistFolderPaths, f.path)
+		}
+	}
+
+	var items []model.ArtworkQueueItem
+
+	albumIDs, err := p.ds.MediaFile(p.ctx).GetAlbumIDsByFolder(lib, folderIDs...)
+	if err != nil {
+		return nil, err
+	}
+	for _, id := range albumIDs {
+		items = append(items, scanArtworkItem(model.KindAlbumArtwork, id))
+	}
+
+	// The resolver climbs to the library root, so the subtree below the folder is the affected set.
+	// A failure here must not discard the album items already collected.
+	artistIDs, err := p.artistIDsUnderPaths(lib, artistFolderPaths)
+	if err != nil {
+		log.Warn(p.ctx, "Scanner: could not map image changes to artists", "lib", lib.Name, err)
+		return items, nil
+	}
+	for _, id := range artistIDs {
+		if id == "" || id == consts.UnknownArtistID || id == consts.VariousArtistsID {
+			continue
+		}
+		items = append(items, scanArtworkItem(model.KindArtistArtwork, id))
+	}
+	return items, nil
+}
+
+func (p *phaseFolders) artistIDsUnderPaths(lib model.Library, paths []string) ([]string, error) {
+	if len(paths) == 0 {
+		return nil, nil
+	}
+	return p.ds.Album(p.ctx).GetSoleAlbumArtistIDsInSubtrees(lib, paths...)
 }
 
 func (p *phaseFolders) finalize(err error) error {
